+++ packages/terra-framework-docs/src/terra-dev-site/test/slide-panel/SlidePanelStart.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import SlidePanel, { SlidePanelPositions } from '../../../SlidePanel';
-import SlidePanelMainContent from '../common/SlidePanelMainContent';
-import SlidePanelPanelContent from '../common/SlidePanelPanelContent';
+import SlidePanel, { SlidePanelPositions } from 'terra-slide-panel';
+import SlidePanelMainContent from './common/SlidePanelMainContent';
+import SlidePanelPanelContent from './common/SlidePanelPanelContent';
 
 const SlidePanelStart = () => (
   <SlidePanel